+++ talisker/context.py
